# openapi/butly.openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "description": "`/api/v1` の共通 error envelope。\n\nHTTP response では常に 4 field すべてを含める（default 省略はしない）。",
        "properties": {
          "code": {
            "description": "frontend 分岐用の安定した error code（snake_case）",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "description": "error code ごとの補足情報（validation error の一覧など）",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "人間可読なエラー説明。secret や内部 path を含めない",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "server log と突合するための request ID（X-Request-ID と同値）",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "details",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "AttachmentSummary": {
        "description": "履歴表示用の添付メタ情報。データ本体は含めない。",
        "properties": {
          "kind": {
            "const": "image",
            "default": "image",
            "title": "Kind",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          }
        },
        "required": [
          "mime_type"
        ],
        "title": "AttachmentSummary",
        "type": "object"
      },
...
      "CitationSource": {
        "description": "Web 検索 / grounding の引用元。provider raw object は正規化して返す。",
        "properties": {
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          }
        },
        "title": "CitationSource",
        "type": "object"
      },
...
      "InstanceListResponse": {
        "description": "`GET /api/v1/instances` の応答。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstanceSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "InstanceListResponse",
        "type": "object"
      },
      "InstanceSummary": {
        "description": "instance 一覧の1件分。",
        "properties": {
          "ai_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Name"
          },
          "has_database": {
            "default": false,
            "title": "Has Database",
            "type": "boolean"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "name"
        ],
        "title": "InstanceSummary",
        "type": "object"
      },
      "Message": {
        "description": "会話履歴の1メッセージ。",
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentSummary"
            },
            "title": "Attachments",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "timezone 付き ISO 8601（UTC 保存）。旧データは None になり得る",
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CitationSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "default": "completed",
            "enum": [
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "text"
        ],
        "title": "Message",
        "type": "object"
      },
      "MessagePage": {
        "description": "`GET /api/v1/instances/{name}/messages` の応答。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "title": "Items",
            "type": "array"
          },
          "last_interaction_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Interaction At"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "さらに古い履歴を取るための cursor。終端は None",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "MessagePage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DesktopToken": {
        "description": "Per-launch desktop token (BUTLY_DESKTOP_TOKEN). Not enforced when the backend runs without a token (dev mode).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Butly backend API. Frontend-facing contract is under /api/v1.",
    "title": "Butly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/instances": {
      "get": {
        "description": "instance の typed 一覧を返す。legacy `GET /instances`（文字列配列）のversioned 置換。破損した instance config があってもその instance はfallback 値で返し、一覧全体は失敗させない。",
        "operationId": "list_instances",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Backend is not ready"
          }
        },
        "security": [
          {
            "DesktopToken": []
          }
        ],
        "summary": "List AI instances",
        "tags": [
          "instances"
        ]
      }
    },
    "/api/v1/instances/{name}/messages": {
      "get": {
        "description": "instance の会話履歴を時系列（古い順）で返す。legacy `GET /history/{name}` のversioned 置換で、timestamp と last_interaction_at を追加する。`before` cursor は予約のみで現時点では無視される（next_cursor も常に null）。",
        "operationId": "list_instance_messages",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "description": "返す最大 message 数",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "返す最大 message 数",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "pagination cursor（予約。現行 storage では未実装のため無視される）",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "pagination cursor（予約。現行 storage では未実装のため無視される）",
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Instance not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Backend is not ready"
          }
        },
        "security": [
          {
            "DesktopToken": []
          }
        ],
        "summary": "Get chat message history",
        "tags": [
          "instances"
        ]
      }
    },
...
  }
}